refactor: ディレクトリ構造を調整
