+++ src/pages/membership/membershipPage.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <h1>This is an membership page</h1>
+  </div>
+</template>
